+++ src/candlewick/utils/MeshData.h
@@ -62,8 +62,8 @@ Mesh convertToMesh(const Device &device, const MeshData &meshData);
 /// \p vertexOffset Vertex buffer offset (in bytes)
 /// \p indexOffset Index buffer offset (in bytes)
 Mesh convertToMesh(const MeshData &meshData, SDL_GPUBuffer *vertexBuffer,
-                   Uint64 vertexOffset, SDL_GPUBuffer *indexBuffer,
-                   Uint64 indexOffset, bool takeOwnership = false);
+                   Uint32 vertexOffset, SDL_GPUBuffer *indexBuffer,
+                   Uint32 indexOffset, bool takeOwnership = false);
 
 /// \brief Upload the contents of a single, individual mesh to the GPU device.
 void uploadMeshToDevice(const Device &device, const Mesh &mesh,
